Treat expired tokens as missing
Sometimes (at least happens on dev machines which go to _sleep_), the tokens are not being automatically refreshed, which causes authentication issues.

This commit checks the `expires` value against the OS time and, if the token (although present on the `TokenStore`) has expired, it is removed and an `:error` is returned (same as if it were missing). This will cause a new token to be requested.