[lib/client] Add ability to retrieve tokens from GCP metadata
Google Cloud Platform has a concept of granting API permissions to instances. This change makes it possible to retrieve tokens from the GCP metadata service in much the same way as the default OAuth service.

Note that token scopes are not currently checked for tokens retrieved from the metadata service.